+++ packages/devtools/src/cjs/index.ts
@@ -1,6 +1,4 @@
-// @ts-expect-error
 import type { Options } from '@wdio/types'
-// @ts-expect-error
 import type { Client, AttachOptions } from '../types'
 
 class Devtools {
@@ -44,4 +42,4 @@ class Devtools {
     }
 }
 
-export = Devtools
+module.exports = Devtools