+++ packages/wdio-allure-reporter/tests/cucumber.suite.test.ts
@@ -139,7 +139,7 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
 
             reporter.onRunnerStart(runnerStart())
             reporter.onSuiteStart(cucumberHelper.featureStart())
-            reporter.onSuiteStart(cucumberHelper.scenarioStart())
+            reporter.onSuiteStart(cucumberHelper.scenarioStart('my-awesome-feature-at-scenario-level'))
             reporter.onHookStart(cucumberHelper.hookStart())
             reporter.onHookEnd(cucumberHelper.hookEnd())
             reporter.onTestStart(cucumberHelper.testStart())
@@ -209,9 +209,11 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
         })
 
         it('should detect analytics labels in test case', () => {
+            const feature = allureResult.labels.find((label: Label) => label.name === LabelName.FEATURE)
             const language = allureResult.labels.find((label: Label) => label.name === LabelName.LANGUAGE)
             const framework = allureResult.labels.find((label: Label) => label.name === LabelName.FRAMEWORK)
 
+            expect(feature.value).toEqual('my-awesome-feature-at-scenario-level')
             expect(language.value).toEqual('javascript')
             expect(framework.value).toEqual('wdio')
         })
@@ -252,7 +254,7 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
             reporter = new AllureReporter({ outputDir, useCucumberStepReporter: true })
 
             reporter.onRunnerStart(runnerStart())
-            reporter.onSuiteStart(cucumberHelper.featureStart())
+            reporter.onSuiteStart(cucumberHelper.featureStart('my-awesome-feature-at-feature-level'))
             reporter.onSuiteStart(cucumberHelper.scenarioStart())
             reporter.onTestStart(cucumberHelper.testStart())
             reporter._consoleOutput = 'some console output'
@@ -276,6 +278,12 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
         afterAll(() => {
             clean(outputDir)
             vi.resetAllMocks()
+        })
+
+        it('should detect analytics labels in test case', () => {
+            expect(allureXml('test-case label[name="feature"]').eq(0).attr('value')).toEqual('my-awesome-feature-at-feature-level')
+            expect(allureXml('test-case label[name="language"]').eq(0).attr('value')).toEqual('javascript')
+            expect(allureXml('test-case label[name="framework"]').eq(0).attr('value')).toEqual('wdio')
         })
 
         it('should report one suite', () => {
@@ -307,8 +315,8 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
             const reporter = new AllureReporter({ outputDir, useCucumberStepReporter: true })
 
             reporter.onRunnerStart(runnerStart())
-            reporter.onSuiteStart(cucumberHelper.featureStart())
-            reporter.onSuiteStart(cucumberHelper.scenarioStart())
+            reporter.onSuiteStart(cucumberHelper.featureStart('my-awesome-feature-at-feature-level'))
+            reporter.onSuiteStart(cucumberHelper.scenarioStart('my-awesome-feature-at-scenario-level'))
             reporter.onTestStart(cucumberHelper.testStart())
             reporter.onTestPass()
             reporter.onTestStart(cucumberHelper.test2start())
@@ -334,6 +342,12 @@ describe('reporter option "useCucumberStepReporter" set to true', () => {
 
         afterAll(() => {
             clean(outputDir)
+        })
+
+        it('should detect analytics labels in test case', () => {
+            expect(allureXml('test-case label[name="feature"]').eq(0).attr('value')).toEqual('my-awesome-feature-at-scenario-level')
+            expect(allureXml('test-case label[name="language"]').eq(0).attr('value')).toEqual('javascript')
+            expect(allureXml('test-case label[name="framework"]').eq(0).attr('value')).toEqual('wdio')
         })
 
         it('should report one suite', () => {